backend/requirements.txt
fastapi
uvicorn
pydantic
pyyaml
paho-mqtt
